Fix: align the GLM-5.3-Flash goldens with their kernel ABIs
Review found nine places where a file's Torch golden and its
@pl.jit.inline signature disagreed. On a scaffold that is the worst
class of defect, because every assignee reads both and the golden is
what makes a row reviewable before the kernel exists.

Two were real bugs rather than mismatches. golden_expert_routed
unsqueezed a route weight that the ABI already declares as
[RECV_DYN, 1], so the expert hidden broadcast across tokens and the
golden returned [RECV_DYN, RECV_DYN, D]. And _round_to_int8 did not
clamp before the int8 cast: harmless on the amax-derived paths, where
the scale bounds the result, but wrong for the static path, where a
recorded input_scale cannot bound an out-of-range activation and 128
wrapped to -128.

The rest close ABI gaps. The KDA and convolution goldens now take
state_rows, so they can model a reordered or paged batch the way the
kernels do. Both sparse-attention goldens take the block table and
request ids, because the index lists carry logical per-request
positions. indexer_topk emits selected_valid beside the selected
pools, which indexer_expand already consumed and nothing produced.
kda_output accumulates in FP32, matching the row-parallel partial its
ABI declares. mla_epilog splits into a W8A8 prefill golden and an
absorbed BF16 decode golden, since one function cannot dequantize an
INT8 weight and pass through a BF16 one. And pool_slots pages at
INDEX_STATE_BLOCK_SIZE rather than BLOCK_SIZE, so pooled row 4 lands
in block 1 offset 0.

Two symbols were doing double duty and are now split. TABLE_DYN meant
both the total blocks in a pool and the per-request block-table width;
the latter is BLOCK_TABLE_DYN. vision_fusion used T_DYN for both the
merged vision rows and the text tokens, which silently asserted they
are equal; the former is VISION_ROWS_DYN.

+++ models/glm5_3_flash/kda_output.py
@@ -32,7 +32,9 @@ def golden_kda_output(
 ) -> torch.Tensor:
     normalized = rms_norm_gated(core_attn_out, norm_weight, out_gate)
     flattened = normalized.reshape(*normalized.shape[:-2], -1)
-    return torch.nn.functional.linear(flattened, w_o)
+    # The kernel emits an FP32 partial for attention_tp to reduce, so the golden
+    # accumulates in FP32 rather than returning the BF16 input dtype.
+    return torch.nn.functional.linear(flattened.float(), w_o.float())
 
 
 @pl.jit.inline
